refactor: move files to components folder
