+++ projects/07-midu-router/src/components/Router.jsx
@@ -1,7 +1,7 @@
-import { EVENTS } from './consts.js'
+import { EVENTS } from '../utils/consts.js'
 import { useState, useEffect, Children } from 'react'
 import { match } from 'path-to-regexp'
-import { getCurrentPath } from './utils.js'
+import { getCurrentPath } from '../utils/getCurrentPath.js'
 
 export function Router ({ children, routes = [], defaultComponent: DefaultComponent = () => <h1>404</h1> }) {
   const [currentPath, setCurrentPath] = useState(getCurrentPath())
